nothing of note

+++ src/test.cpp
@@ -19,7 +19,7 @@ int main(void){
     //inf_int<int8_t> a = 1000;
 
     inf_int<int8_t> a = 20;
-    inf_int<int8_t> b = 40;
+    inf_int<unsigned int> b = 40;
 
 
     a+=b;
